Move signer back to being protected
